# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Pill Guardian API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ErrorResponse {
  error: string;
}

export interface SuccessResponse {
  message: string;
}

export interface Guardian {
  id: number;
  name: string;
  nickname: string;
  phone: string;
  createdAt: string;
  updatedAt: string;
}

export interface GuardianInput {
  name: string;
  nickname: string;
  phone: string;
}

export interface Patient {
  id: number;
  name: string;
  nickname: string;
  phone: string;
  createdAt: string;
  updatedAt: string;
}

export interface PatientInput {
  name: string;
  nickname: string;
  phone: string;
}

export interface Pill {
  id: number;
  name: string;
  description: string;
  steps: string;
  /** Time in HH:MM format */
  scheduledTime: string;
  done: boolean;
  createdAt: string;
  updatedAt: string;
}

export interface PillInput {
  name: string;
  description: string;
  steps: string;
  /** Time in HH:MM format */
  scheduledTime: string;
}

export interface Esp32Status {
  connected: boolean;
  connectedAt?: string | null;
}