# lib/api-zod/src/generated/types/esp32Status.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Pill Guardian API
 * OpenAPI spec version: 0.1.0
 */

export interface Esp32Status {
  connected: boolean;
  connectedAt?: Date | null;
}