# lib/api-zod/src/generated/types/guardian.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Pill Guardian API
 * OpenAPI spec version: 0.1.0
 */

export interface Guardian {
  id: number;
  name: string;
  nickname: string;
  phone: string;
  createdAt: Date;
  updatedAt: Date;
}