# lib/api-zod/src/generated/types/guardianInput.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Pill Guardian API
 * OpenAPI spec version: 0.1.0
 */

export interface GuardianInput {
  name: string;
  nickname: string;
  phone: string;
}