# lib/api-zod/src/generated/types/pillInput.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Pill Guardian API
 * OpenAPI spec version: 0.1.0
 */

export interface PillInput {
  name: string;
  description: string;
  steps: string;
  /** Time in HH:MM format */
  scheduledTime: string;
  /** Comma-separated day abbreviations e.g. "Mon,Wed,Fri" */
  daysOfWeek: string;
}